VY: Modified make_check.sh with colors.

+++ src/ELSI/make_check.sh
@@ -7,56 +7,83 @@ set -e # Stop on error
 echo
 ${MPI_EXEC} -n 4 ./test_ev_real.x ${TOMATO_SEED} 1 > ev_real_elpa.log
 if (! grep -q "Passed" <./ev_real_elpa.log); then
+   tput setaf 5
    echo "FAILED:  elsi_ev_real + ELPA (mp)"
+   tput sgr0
    echo "See `pwd`/ev_real_elpa.log for details."
 else
+   tput setaf 10
    echo "PASSED:  elsi_ev_real + ELPA (mp)"
+   tput sgr0
    rm ev_real_elpa.log
 fi
 
-echo
-${MPI_EXEC} -n 1 ./test_ev_real.x ${TOMATO_SEED} 1 > ev_real_elpa_sp.log
-if (! grep -q "Passed" <./ev_real_elpa_sp.log); then
-   echo "FAILED:  elsi_ev_real + ELPA (sp)"
-   echo "See `pwd`/ev_real_elpa_sp.log for details."
-   echo "- If you are using the GPU-accelerated version of ELPA and the log file contains the error"
-   echo "- 'ELPA GPU version needs blocksize = 128', you may safely ignore the failure of this test."
-else
-   echo "PASSED:  elsi_ev_real + ELPA (sp)"
-   rm ev_real_elpa_sp.log
+if [ "$ELPA2_KERNEL" != "GPU" ]
+then
+   echo
+   ${MPI_EXEC} -n 1 ./test_ev_real.x ${TOMATO_SEED} 1 > ev_real_elpa_sp.log
+   if (! grep -q "Passed" <./ev_real_elpa_sp.log); then
+      tput setaf 5
+      echo "FAILED:  elsi_ev_real + ELPA (sp)"
+      tput sgr0
+      echo "See `pwd`/ev_real_elpa_sp.log for details."
+   else
+      tput setaf 10
+      echo "PASSED:  elsi_ev_real + ELPA (sp)"
+      tput sgr0
+      rm ev_real_elpa_sp.log
+   fi
 fi
 
 echo
 ${MPI_EXEC} -n 4 ./test_dm_real.x ${TOMATO_SEED} 1 > dm_real_elpa.log
 if (! grep -q "Passed" <./dm_real_elpa.log); then
+   tput setaf 5
    echo "FAILED:  elsi_dm_real + ELPA"
+   tput sgr0
    echo "See `pwd`/dm_real_elpa.log for details."
 else
+   tput setaf 10
    echo "PASSED:  elsi_dm_real + ELPA"
+   tput sgr0
    rm dm_real_elpa.log
 fi
 
 echo
 ${MPI_EXEC} -n 4 ./test_dm_real.x ${TOMATO_SEED} 2 > dm_real_libomm.log
 if (! grep -q "Passed" <./dm_real_libomm.log); then
+   tput setaf 5
    echo "FAILED:  elsi_dm_real + libOMM"
+   tput sgr0
    echo "See `pwd`/dm_real_libomm.log for details."
 else
+   tput setaf 10
    echo "PASSED:  elsi_dm_real + libOMM"
+   tput sgr0
    rm dm_real_libomm.log libOMM.log
 fi
 
 echo
-${MPI_EXEC} -n 4 ./test_dm_real.x ${TOMATO_SEED} 3 > dm_real_pexsi.log
-if (! grep -q "Passed" <./dm_real_pexsi.log); then
-   echo "FAILED:  elsi_dm_real + PEXSI"
-   echo "See `pwd`/dm_real_pexsi.log for details."
-   echo "- If you compiled ELSI without C++ support, PEXSI is disabled, and this failure is expected."
-   echo "- In this case, please recompile ELSI with C++ support to enable PEXSI, or ignore this error"
-   echo "- message to continue using ELSI without PEXSI."
+if [ "$DISABLE_CXX" != "yes" ]
+then
+   ${MPI_EXEC} -n 4 ./test_dm_real.x ${TOMATO_SEED} 3 > dm_real_pexsi.log
+   if (! grep -q "Passed" <./dm_real_pexsi.log); then
+      tput setaf 5
+      echo "FAILED:  elsi_dm_real + PEXSI"
+      tput sgr0
+      echo "See `pwd`/dm_real_pexsi.log for details."
+   else
+      tput setaf 10
+      echo "PASSED:  elsi_dm_real + PEXSI"
+      tput sgr0
+      rm dm_real_pexsi.log logPEXSI0
+   fi
 else
-   echo "PASSED:  elsi_dm_real + PEXSI"
-   rm dm_real_pexsi.log logPEXSI0
+   tput setaf 5
+   echo "PEXSI has been disabled because ELSI is compiled without C++ support."
+   echo "Please recompile ELSI with C++ support to enable PEXSI, or continue"
+   echo "using ELSI without PEXSI."
+   tput sgr0
 fi
 
 echo
